requirements.txt
flask==2.0.3
dnspython==2.2.1
werkzeug==2.0.3
python-dotenv==0.19.2  